+++ Validators/DatiRiepilogoValidator.cs
@@ -26,7 +26,8 @@ namespace FatturaElettronica.Validators
             RuleFor(x => x.Natura)
                 .Must(natura => natura != "N6")
                 .When(x => x.EsigibilitaIVA == "S")
-                .WithMessage("Natura con valore 'N6' (inversion contabile) a fronte EsigiblitaIVA uguale a 'S' (scission pagamenti)")
+                .WithMessage(
+                    "Natura con valore 'N6' (inversion contabile) a fronte EsigiblitaIVA uguale a 'S' (scission pagamenti)")
                 .WithErrorCode("00420");
             RuleFor(x => x.Imposta)
                 .Must((challenge, _) => ImpostaValidateAgainstError00421(challenge))
@@ -38,6 +39,9 @@ namespace FatturaElettronica.Validators
                 .Length(1, 100)
                 .Latin1SupplementValidator()
                 .When(x => !string.IsNullOrEmpty(x.RiferimentoNormativo));
+            RuleFor(x => x.RiferimentoNormativo)
+                .Must(riferimento => !string.IsNullOrEmpty(riferimento))
+                .When(x => !string.IsNullOrEmpty(x.Natura));
             RuleFor(x => x.SpeseAccessorie)
                 .ScalePrecision2DecimalType();
             RuleFor(x => x.ImponibileImporto)
@@ -47,9 +51,12 @@ namespace FatturaElettronica.Validators
             RuleFor(x => x.Arrotondamento)
                 .ScalePrecision8DecimalType();
         }
-        private bool ImpostaValidateAgainstError00421(DatiRiepilogo datiRiepilogo)
+
+        private static bool ImpostaValidateAgainstError00421(DatiRiepilogo datiRiepilogo)
         {
-            return (Math.Abs(datiRiepilogo.Imposta - decimal.Parse(((datiRiepilogo.AliquotaIVA * datiRiepilogo.ImponibileImporto) / 100).ToString("0.00"))) <= 0.01m);
+            return (Math.Abs(datiRiepilogo.Imposta -
+                             decimal.Parse(((datiRiepilogo.AliquotaIVA * datiRiepilogo.ImponibileImporto) / 100)
+                                 .ToString("0.00"))) <= 0.01m);
         }
     }
 }